+++ crates/protocol-tokio/src/host.rs
@@ -18,25 +18,26 @@ use std::path::Path;
 use anyhow::Result;
 use tokio::io::{AsyncRead, AsyncWrite};
 use tokio::net::{TcpStream, UnixStream};
+use tokio::sync::Mutex;
 use wasefire_protocol::DynFuture;
 
 use crate::common::{read, write};
 
 /// Holds a connection to a device.
 pub struct Connection<S: AsyncRead + AsyncWrite + Send + Unpin> {
-    stream: S,
+    stream: Mutex<S>,
 }
 
 impl Connection<UnixStream> {
     pub async fn new_unix(path: &Path) -> Result<Self> {
-        let stream = UnixStream::connect(path).await?;
+        let stream = Mutex::new(UnixStream::connect(path).await?);
         Ok(Connection { stream })
     }
 }
 
 impl Connection<TcpStream> {
     pub async fn new_tcp(addr: SocketAddr) -> Result<Self> {
-        let stream = TcpStream::connect(addr).await?;
+        let stream = Mutex::new(TcpStream::connect(addr).await?);
         Ok(Connection { stream })
     }
 }
@@ -44,11 +45,11 @@ impl Connection<TcpStream> {
 impl<T: AsyncRead + AsyncWrite + Send + Unpin + 'static> wasefire_protocol::Connection
     for Connection<T>
 {
-    fn read(&mut self) -> DynFuture<'_, Box<[u8]>> {
-        Box::pin(async move { Ok(read(&mut self.stream).await?) })
+    fn write<'a>(&'a self, request: &'a [u8]) -> DynFuture<'a, ()> {
+        Box::pin(async move { Ok(write(&mut *self.stream.lock().await, request).await?) })
     }
 
-    fn write<'a>(&'a mut self, request: &'a [u8]) -> DynFuture<'a, ()> {
-        Box::pin(async move { Ok(write(&mut self.stream, request).await?) })
+    fn read(&self) -> DynFuture<'_, Box<[u8]>> {
+        Box::pin(async move { Ok(read(&mut *self.stream.lock().await).await?) })
     }
 }